@@ -19,48 +26,54 @@
     "icon": "ng g ng-alain:plugin icon",
     "new-dev": "node scripts/new-dev.js"
   },
-  "private": true,
   "dependencies": {
-    "@angular/animations": "~8.1.2",
-    "@angular/common": "~8.1.2",
-    "@angular/compiler": "~8.1.2",
-    "@angular/core": "~8.1.2",
-    "@angular/forms": "~8.1.2",
-    "@angular/platform-browser": "~8.1.2",
-    "@angular/platform-browser-dynamic": "~8.1.2",
-    "@angular/router": "~8.1.2",
-    "ng-alain": "^8.2.0",
+    "@angular/animations": "~8.2.3",
+    "@angular/common": "~8.2.3",
+    "@angular/compiler": "~8.2.3",
+    "@angular/core": "~8.2.3",
+    "@angular/forms": "~8.2.3",
+    "@angular/platform-browser": "~8.2.3",
+    "@angular/platform-browser-dynamic": "~8.2.3",
+    "@angular/router": "~8.2.3",
     "rxjs": "~6.4.0",
-    "tslib": "^1.9.0",
+    "tslib": "^1.10.0",
     "zone.js": "~0.9.1",
-    "ng-zorro-antd": "^8.1.0",
-    "screenfull": "^4.2.0",
-    "ajv": "^6.10.0",
-    "@delon/abc": "^8.2.0",
-    "@delon/acl": "^8.2.0",
-    "@delon/auth": "^8.2.0",
-    "@delon/cache": "^8.2.0",
-    "@delon/form": "^8.2.0",
-    "@delon/mock": "^8.2.0",
-    "@delon/theme": "^8.2.0",
-    "@delon/util": "^8.2.0",
-    "@delon/chart": "^8.2.0",
-    "@ngx-translate/core": "^11.0.1",
-    "@ngx-translate/http-loader": "^4.0.0",
     "@antv/data-set": "^0.10.2",
     "@antv/g2": "^3.5.3",
     "@antv/g2-plugin-slider": "^2.1.1",
-    "@angular/elements": "~8.1.2",
-    "viser-ng": "2.4.4",
-    "oidc-client": "1.7.0",
-    "moment": "2.24.0",
-    "lodash": "4.17.13"
+    "@ngx-translate/core": "^11.0.1",
+    "@ngx-translate/http-loader": "^4.0.0",
+    "ajv": "^6.10.2",
+    "@delon/abc": "^8.5.0",
+    "@delon/acl": "^8.5.0",
+    "@delon/auth": "^8.5.0",
+    "@delon/cache": "^8.5.0",
+    "@delon/chart": "^8.5.0",
+    "@delon/form": "^8.5.0",
+    "@delon/mock": "^8.5.0",
+    "@delon/theme": "^8.5.0",
+    "@delon/util": "^8.5.0",
+    "file-saver": "^2.0.2",
+    "ng-zorro-antd": "^8.3.1",
+    "screenfull": "^4.2.1",
+    "qrious": "^4.0.2",
+    "quill": "^1.3.6",
+    "quill-image-resize-module": "^3.0.0",
+    "perfect-scrollbar": "^1.4.0",
+    "ngx-image-gallery": "^1.3.1",
+    "masonry-layout": "^4.2.2",
+    "@agm/core": "^1.0.0-beta.5",
+    "angular-baidu-maps": "^3.0.0",
+    "@swimlane/ngx-charts": "^12.0.1",
+    "ngx-trend": "^4.0.0",
+    "element-closest": "^3.0.1",
+    "js-base64": "^2.5.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.801.2",
-    "@angular/cli": "~8.1.2",
-    "@angular/compiler-cli": "~8.1.2",
-    "@angular/language-service": "~8.1.2",
+    "@angular-devkit/build-angular": "~0.803.14",
+    "@angular/cli": "~8.3.1",
+    "@angular/compiler-cli": "~8.2.3",
+    "@angular/language-service": "~8.2.3",
     "@types/node": "~8.9.4",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
@@ -75,16 +88,17 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.15.0",
-    "typescript": "~3.4.3",
-    "ng-alain": "^8.2.0",
-    "ng-alain-codelyzer": "^0.0.1",
-    "@delon/testing": "^8.2.0",
-    "antd-theme-generator": "^1.1.6",
-    "less-bundle-promise": "^1.0.7",
+    "typescript": "~3.5.3",
+    "@angularclass/hmr": "^2.1.3",
+    "@types/jszip": "^3.1.6",
+    "@types/mockjs": "^1.0.2",
+    "codecov": "^3.5.0",
     "tslint-config-prettier": "^1.18.0",
-    "tslint-language-service": "^0.9.9",
+    "husky": "^3.0.4",
+    "gh-pages": "^2.1.1",
     "lint-staged": "^8.2.1",
-    "husky": "^3.0.0",
+    "less-bundle-promise": "^1.0.7",
+    "mockjs": "^1.0.1-beta3",
     "prettier": "^1.18.2",
     "prettier-stylelint": "^0.4.2",
     "stylelint": "^10.1.0",
