@@ -1,25 +1,24 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-  import { <%= componentName %> } from './<%= dasherize(name) %>.component';
+import { <%= componentName %> } from './<%= dasherize(name) %>.component';
 
-  describe('<%= componentName %>', () => {
-    let component: <%= componentName %>;
-    let fixture: ComponentFixture<<%= componentName %>>;
+describe('<%= componentName %>', () => {
+  let component: <%= componentName %>;
+  let fixture: ComponentFixture<<%= componentName %>>;
 
-    beforeEach(async(() => {
-      TestBed.configureTestingModule({
-        declarations: [ <%= componentName %> ]
-      })
-      .compileComponents();
-    }));
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ <%= componentName %> ]
+    })
+    .compileComponents();
+  }));
 
-    beforeEach(() => {
-      fixture = TestBed.createComponent(<%= componentName %>);
-      component = fixture.componentInstance;
-      fixture.detectChanges();
-    });
-
-    it('should create', () => {
-      expect(component).toBeTruthy();
-    });
+  beforeEach(() => {
+    fixture = TestBed.createComponent(<%= componentName %>);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
   });
-  
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
