@@ -6,15 +6,15 @@ import {
   HttpHandler,
   HttpErrorResponse,
   HttpEvent,
-  HttpResponse,
   HttpResponseBase,
 } from '@angular/common/http';
 import { Observable, of, throwError } from 'rxjs';
 import { mergeMap, catchError } from 'rxjs/operators';
 import { NzMessageService, NzNotificationService } from 'ng-zorro-antd';
-import { _HttpClient } from '@delon/theme';
+import { _HttpClient, ALAIN_I18N_TOKEN } from '@delon/theme';
 import { environment } from '@env/environment';
 import { DA_SERVICE_TOKEN, ITokenService } from '@delon/auth';
+import { I18NService } from '../i18n/i18n.service';
 
 const CODEMESSAGE = {
   200: '服务器成功返回请求的数据。',
@@ -41,8 +41,12 @@ const CODEMESSAGE = {
 export class DefaultInterceptor implements HttpInterceptor {
   constructor(private injector: Injector) {}
 
-  private get notification(): NzNotificationService {
-    return this.injector.get(NzNotificationService);
+  get msg(): NzMessageService {
+    return this.injector.get(NzMessageService);
+  }
+
+  private get lang(): string {
+    return (this.injector.get(ALAIN_I18N_TOKEN) as I18NService).currentLang;
   }
 
   private goTo(url: string) {
@@ -50,12 +54,10 @@ export class DefaultInterceptor implements HttpInterceptor {
   }
 
   private checkStatus(ev: HttpResponseBase) {
-    if ((ev.status >= 200 && ev.status < 300) || ev.status === 401) {
-      return;
-    }
+    if (ev.status >= 200 && ev.status < 300) return;
 
     const errortext = CODEMESSAGE[ev.status] || ev.statusText;
-    this.notification.error(`请求错误 ${ev.status}: ${ev.url}`, errortext);
+    this.injector.get(NzNotificationService).error(`请求错误 ${ev.status}: ${ev.url}`, errortext);
   }
 
   private handleData(ev: HttpResponseBase): Observable<any> {
@@ -72,26 +74,23 @@ export class DefaultInterceptor implements HttpInterceptor {
         //  错误内容：{ status: 1, msg: '非法参数' }
         //  正确内容：{ status: 0, response: {  } }
         // 则以下代码片断可直接适用
-        if (ev instanceof HttpResponse) {
-          const body: any = ev.body;
-          if (body && body.isSucceed !== undefined && body.entity !== undefined && body.statusCode !== undefined) {
-            if (body && body.statusCode !== 200) {
-              this.notification.error(`请求错误 ${ev.status}: ${ev.url}`, body.message);
-              // 继续抛出错误中断后续所有 Pipe、subscribe 操作，因此：
-              // this.http.get('/').subscribe() 并不会触发
-              return throwError({});
-            } else {
-              // 重新修改 `body` 内容为 `response` 内容，对于绝大多数场景已经无须再关心业务状态码
-              return of(new HttpResponse(Object.assign(event, { body: body.entity })));
-            }
-          }
-          // 或者依然保持完整的格式
-          return of(ev);
-        }
+        // if (event instanceof HttpResponse) {
+        //     const body: any = event.body;
+        //     if (body && body.status !== 0) {
+        //         this.msg.error(body.msg);
+        //         // 继续抛出错误中断后续所有 Pipe、subscribe 操作，因此：
+        //         // this.http.get('/').subscribe() 并不会触发
+        //         return throwError({});
+        //     } else {
+        //         // 重新修改 `body` 内容为 `response` 内容，对于绝大多数场景已经无须再关心业务状态码
+        //         return of(new HttpResponse(Object.assign(event, { body: body.response })));
+        //         // 或者依然保持完整的格式
+        //         return of(event);
+        //     }
+        // }
         break;
-      case 401:
-        this.notification.error(`未登录或登录已过期，请重新登录。`, ``);
-        // 清空 token 信息
+      case 401: // 未登录状态码
+        // 请求错误 401: https://preview.pro.ant.design/api/401 用户没有权限（令牌、用户名、密码错误）。
         (this.injector.get(DA_SERVICE_TOKEN) as ITokenService).clear();
         this.goTo('/passport/login');
         break;
@@ -118,6 +117,7 @@ export class DefaultInterceptor implements HttpInterceptor {
     }
 
     const newReq = req.clone({ url });
+    newReq.headers.set('lang', this.lang);
     return next.handle(newReq).pipe(
       mergeMap((event: any) => {
         // 允许统一对请求错误处理
