@@ -1,85 +1,74 @@
-import { Injectable, Injector, Inject } from '@angular/core';
-import { Router } from '@angular/router';
-import { HttpClient } from '@angular/common/http';
+import { Injectable, Inject } from '@angular/core';
 import { DOCUMENT } from '@angular/common';
+import { HttpClient } from '@angular/common/http';
 import { zip } from 'rxjs';
 import { catchError } from 'rxjs/operators';
-import { MenuService, SettingsService, TitleService, ALAIN_I18N_TOKEN } from '@delon/theme';
-import { DA_SERVICE_TOKEN, ITokenService } from '@delon/auth';
-import { ACLService } from '@delon/acl';
-import { TranslateService } from '@ngx-translate/core';
-import { I18NService } from '../i18n/i18n.service';
 
-import { NzIconService } from 'ng-zorro-antd/icon';
+import { SettingsService, ALAIN_I18N_TOKEN, TitleService } from '@delon/theme';
+import { ACLService } from '@delon/acl';
+import { NzIconService } from 'ng-zorro-antd';
+
 import { ICONS_AUTO } from '../../../style-icons-auto';
 import { ICONS } from '../../../style-icons';
 
+import { I18NService } from '@core/i18n/i18n.service';
+
 /**
- * Used for application startup
- * Generally used to get the basic data of the application, like: Menu Data, User Data, etc.
+ * 用于应用启动时
+ * 一般用来获取应用所需要的基础数据等
  */
 @Injectable()
 export class StartupService {
   constructor(
     iconSrv: NzIconService,
-    private menuService: MenuService,
-    private translate: TranslateService,
-    @Inject(ALAIN_I18N_TOKEN) private i18n: I18NService,
     private settingService: SettingsService,
     private aclService: ACLService,
-    private titleService: TitleService,
-    @Inject(DA_SERVICE_TOKEN) private tokenService: ITokenService,
     private httpClient: HttpClient,
     @Inject(DOCUMENT) private doc: any,
-    private injector: Injector,
+    @Inject(ALAIN_I18N_TOKEN) private i18n: I18NService,
+    private titleSrv: TitleService,
   ) {
     iconSrv.addIcon(...ICONS_AUTO, ...ICONS);
-  }
-
-  private getAppData(resolve: any, reject: any) {
-    zip(
-      //this.httpClient.get(`${window.location.origin}/assets/tmp/i18n/${this.i18n.defaultLang}.json`),
-      //this.httpClient.get(`${window.location.origin}/assets/tmp/app-data.json`),
-      this.httpClient.get(`http://localhost:280/api/language/getlangdataasync?lang=${this.i18n.defaultLang}`, { headers: { "Content-Type": "application/json" } }),
-      this.httpClient.get(`http://localhost:280/api/startup/getappdataasync`, { headers: { "Content-Type": "application/json" } })
-    )
-      .pipe(
-        catchError(([langData, appData]) => {
-          resolve(null);
-          return [langData, appData];
-        }),
-      )
-      .subscribe(
-        ([langData, appData]) => {
-          // Setting language data
-          this.translate.setTranslation(this.i18n.defaultLang, langData);
-          this.translate.setDefaultLang(this.i18n.defaultLang);
-          // Application data
-          const res: any = appData;
-          // Application information: including site name, description, year
-          this.settingService.setApp(res.app);
-          // User information: including name, avatar, email address
-          this.settingService.setUser(res.user);
-          // ACL: Set the permissions to full, https://ng-alain.com/acl/getting-started
-          this.aclService.setFull(true);
-          // Menu data, https://ng-alain.com/theme/menu
-          this.menuService.add(res.menu);
-          // Can be set page suffix title, https://ng-alain.com/theme/title
-          this.titleService.suffix = res.app.name;
-        },
-        () => {},
-        () => {
-          resolve(null);
-          this.removeLoading();
-        },
-      );
   }
 
   load(): Promise<any> {
     // only works with promises
     // https://github.com/angular/angular/issues/15088
-    return new Promise((resolve, reject) => {
-      this.getAppData(resolve, reject);
+    return new Promise(resolve => {
+      zip(
+        this.httpClient.get(`assets/tmp/i18n/${this.i18n.currentLang}.json`),
+        this.httpClient.get(`assets/tmp/app-data.json`),
+      )
+        .pipe(
+          // 接收其他拦截器后产生的异常消息
+          catchError(([langData, appData]) => {
+            resolve(null);
+            return [langData, appData];
+          }),
+        )
+        .subscribe(
+          ([langData, appData]) => {
+            // Load general language data
+            this.i18n.load(this.i18n.currentLang, langData);
+
+            // application data
+            const res: any = appData;
+            // 应用信息：包括站点名、描述、年份
+            this.settingService.setApp(res.app);
+            // 用户信息：包括姓名、头像、邮箱地址
+            this.settingService.setUser(res.user);
+            // ACL：设置权限为全量
+            this.aclService.setFull(true);
+            // 设置标题
+            this.titleSrv.default = '';
+            this.titleSrv.suffix = 'Ng Alain MS';
+          },
+          () => {},
+          () => {
+            resolve(null);
+            this.removeLoading();
+          },
+        );
     });
   }
 
