@@ -6,8 +6,9 @@ import { ExceptionRoutingModule } from './exception-routing.module';
 import { Exception403Component } from './403.component';
 import { Exception404Component } from './404.component';
 import { Exception500Component } from './500.component';
+import { ExceptionTriggerComponent } from './trigger.component';
 
-const COMPONENTS = [Exception403Component, Exception404Component, Exception500Component];
+const COMPONENTS = [Exception403Component, Exception404Component, Exception500Component, ExceptionTriggerComponent];
 const COMPONENTS_NOROUNT = [];
 
 @NgModule({
@@ -15,4 +16,4 @@ const COMPONENTS_NOROUNT = [];
   declarations: [...COMPONENTS, ...COMPONENTS_NOROUNT],
   entryComponents: COMPONENTS_NOROUNT,
 })
-export class ExceptionModule { }
+export class ExceptionModule {}
