@@ -1,36 +1,81 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { environment } from '@env/environment';
-
-import { OidcAuthorize as Authorize } from '@util';
+import { SimpleGuard } from '@delon/auth';
+// layout
 import { MSLayoutComponent } from '@brand';
-import { LayoutFullScreenComponent } from '../layout/fullscreen/fullscreen.component';
-
-import { DashboardComponent } from './dashboard/dashboard.component';
+// single pages
 import { CallbackComponent } from './callback/callback.component';
 
 const routes: Routes = [
+  // Full pages
   {
     path: '',
     component: MSLayoutComponent,
-     canActivate: [Authorize],
-     canActivateChild: [Authorize],
+    canActivate: [SimpleGuard],
     children: [
-      { path: '', redirectTo: 'dashboard', pathMatch: 'full' },
-      { path: 'dashboard', component: DashboardComponent, data: { title: '仪表盘' } },
-      { path: 'exception', loadChildren: () => import('./exception/exception.module').then(m => m.ExceptionModule) },
-      // 业务子模块
-      { path: 'api', loadChildren: () => import('./api/api.module').then(m => m.APIModule) },
+      { path: '', redirectTo: 'home', pathMatch: 'full' },
+      { path: 'home', loadChildren: './home/home.module#HomeModule' },
+      {
+        path: 'component',
+        loadChildren: './component/component.module#ComponentModule',
+      },
+      {
+        path: 'dns',
+        loadChildren: './dns/dns.module#DnsModule',
+      },
+      {
+        path: 'api',
+        loadChildren: './api/api.module#APIModule',
+      },
+      {
+        path: 'authorize',
+        loadChildren: './authorize/authorize.module#AuthorizeModule',
+      },
+      {
+        path: 'user',
+        loadChildren: './user/user.module#UserModule',
+      },
+      {
+        path: 'notifications',
+        loadChildren: './notifications/notifications.module#NotificationsModule',
+      },
+      // Exception
+      {
+        path: 'exception',
+        loadChildren: './exception/exception.module#ExceptionModule',
+      },
     ],
   },
-  // 全屏布局
+  // Single pages (Not sidebar)
   {
-    path: 'fullscreen',
-    component: LayoutFullScreenComponent,
-    children: [],
+    path: '',
+    component: MSLayoutComponent,
+    canActivate: [SimpleGuard],
+    data: { hasAllNav: true, hasSidebar: false },
+    children: [
+      { path: 'smart', loadChildren: './smart/smart.module#SmartModule' },
+      { path: 'data', loadChildren: './data/data.module#DataModule' },
+      {
+        path: 'finance',
+        loadChildren: './finance/finance.module#FinanceModule',
+      },
+    ],
   },
-  // 单页不包裹
-  { path: 'callback', component: CallbackComponent },
+  // Help Document
+  {
+    path: 'help',
+    component: MSLayoutComponent,
+    data: { hasAllNav: true, hasSidebar: false },
+    children: [{ path: '', loadChildren: './help/help.module#HelpModule' }],
+  },
+  // account
+  {
+    path: 'account',
+    loadChildren: './account/account.module#AccountModule',
+  },
+  // 单页不包裹Layout
+  { path: 'callback/:type', component: CallbackComponent },
   { path: '**', redirectTo: 'exception/404' },
 ];
 
@@ -38,8 +83,6 @@ const routes: Routes = [
   imports: [
     RouterModule.forRoot(routes, {
       useHash: environment.useHash,
-      // NOTICE: If you use `reuse-tab` component and turn on keepingScroll you can set to `disabled`
-      // Pls refer to https://ng-alain.com/components/reuse-tab
       scrollPositionRestoration: 'top',
     }),
   ],
