@@ -2,22 +2,18 @@ import { NgModule } from '@angular/core';
 
 import { SharedModule } from '@shared';
 import { RouteRoutingModule } from './routes-routing.module';
-
-import { DashboardComponent } from './dashboard/dashboard.component';
+// single pages
 import { CallbackComponent } from './callback/callback.component';
 
 const COMPONENTS = [
-  DashboardComponent,
+  // single pages
   CallbackComponent,
 ];
 const COMPONENTS_NOROUNT = [];
 
 @NgModule({
   imports: [SharedModule, RouteRoutingModule],
-  declarations: [
-    ...COMPONENTS,
-    ...COMPONENTS_NOROUNT
-  ],
-  entryComponents: COMPONENTS_NOROUNT
+  declarations: [...COMPONENTS, ...COMPONENTS_NOROUNT],
+  entryComponents: COMPONENTS_NOROUNT,
 })
-export class RoutesModule { }
+export class RoutesModule {}
