@@ -1,29 +1,27 @@
 import { NgModule } from '@angular/core';
-import { DelonFormModule, WidgetRegistry } from '@delon/form';
 import { SharedModule } from '@shared';
+import { DelonFormModule, WidgetRegistry } from '@delon/form';
 
-// import { TinymceWidget } from './widgets/tinymce/tinymce.widget';
-// import { UEditorWidget } from './widgets/ueditor/ueditor.widget';
+import { EditorWidget } from './widgets/editor/editor.widget';
+import { ImgWidget } from './widgets/img/img.widget';
+import { AddressWidget } from './widgets/address/address.widget';
 
 export const SCHEMA_THIRDS_COMPONENTS = [
-  // TinymceWidget,
-  // UEditorWidget
+  EditorWidget,
+  ImgWidget,
+  AddressWidget,
 ];
 
 @NgModule({
   declarations: SCHEMA_THIRDS_COMPONENTS,
   entryComponents: SCHEMA_THIRDS_COMPONENTS,
-  imports: [
-    SharedModule,
-    DelonFormModule.forRoot()
-  ],
-  exports: [
-    ...SCHEMA_THIRDS_COMPONENTS
-  ]
+  imports: [SharedModule, DelonFormModule.forRoot()],
+  exports: [...SCHEMA_THIRDS_COMPONENTS],
 })
 export class JsonSchemaModule {
   constructor(widgetRegistry: WidgetRegistry) {
-    // widgetRegistry.register(TinymceWidget.KEY, TinymceWidget);
-    // widgetRegistry.register(UEditorWidget.KEY, UEditorWidget);
+    widgetRegistry.register(EditorWidget.KEY, EditorWidget);
+    widgetRegistry.register(ImgWidget.KEY, ImgWidget);
+    widgetRegistry.register(AddressWidget.KEY, AddressWidget);
   }
 }
