@@ -5,7 +5,7 @@
 export const environment = {
   SERVER_URL: `./`,
   production: false,
-  useHash: false,
+  useHash: true,
   hmr: true,
 };
 
