@@ -1,102 +1,100 @@
 /*
-* Automatically generated by 'ng g ng-alain:plugin icon'
-* @see https://ng-alain.com/cli/plugin#icon
-*/
+ * Automatically generated by 'ng g ng-alain:plugin icon'
+ * @see https://ng-alain.com/cli/plugin#icon
+ */
 
 import {
   AlipayCircleOutline,
-  ApiOutline,
   AppstoreOutline,
-  ArrowDownOutline,
   BookOutline,
-  CloudOutline,
-  CopyrightOutline,
+  BranchesOutline,
+  CaretRightOutline,
   CustomerServiceOutline,
-  DashboardOutline,
-  DatabaseOutline,
-  DingdingOutline,
-  DislikeOutline,
+  DesktopOutline,
+  DollarOutline,
   DownloadOutline,
-  ForkOutline,
-  FrownOutline,
-  FullscreenExitOutline,
-  FullscreenOutline,
+  EditOutline,
+  FacebookOutline,
+  FileSearchOutline,
+  FolderOutline,
+  GiftOutline,
   GithubOutline,
   GlobalOutline,
-  HddOutline,
-  LaptopOutline,
-  LikeOutline,
+  HeartFill,
   LockOutline,
-  LogoutOutline,
   MailOutline,
   MenuFoldOutline,
-  MenuUnfoldOutline,
-  MessageOutline,
-  PayCircleOutline,
+  PhoneOutline,
   PieChartOutline,
-  PrinterOutline,
-  RocketOutline,
+  PlusCircleOutline,
+  QrcodeOutline,
+  QuestionCircleFill,
+  QuestionOutline,
+  RollbackOutline,
+  SafetyOutline,
   ScanOutline,
   SettingOutline,
   ShareAltOutline,
   ShoppingCartOutline,
-  SoundOutline,
+  StarFill,
   StarOutline,
+  SyncOutline,
+  TableOutline,
   TaobaoCircleOutline,
-  TaobaoOutline,
-  TeamOutline,
+  ToTopOutline,
   ToolOutline,
   TrophyOutline,
-  UsbOutline,
   UserOutline,
-  WeiboCircleOutline
+  WechatOutline,
+  WeiboCircleOutline,
+  WeiboOutline,
+  ZhihuOutline,
 } from '@ant-design/icons-angular/icons';
 
 export const ICONS_AUTO = [
   AlipayCircleOutline,
-  ApiOutline,
   AppstoreOutline,
-  ArrowDownOutline,
   BookOutline,
-  CloudOutline,
-  CopyrightOutline,
+  BranchesOutline,
+  CaretRightOutline,
   CustomerServiceOutline,
-  DashboardOutline,
-  DatabaseOutline,
-  DingdingOutline,
-  DislikeOutline,
+  DesktopOutline,
+  DollarOutline,
   DownloadOutline,
-  ForkOutline,
-  FrownOutline,
-  FullscreenExitOutline,
-  FullscreenOutline,
+  EditOutline,
+  FacebookOutline,
+  FileSearchOutline,
+  FolderOutline,
+  GiftOutline,
   GithubOutline,
   GlobalOutline,
-  HddOutline,
-  LaptopOutline,
-  LikeOutline,
+  HeartFill,
   LockOutline,
-  LogoutOutline,
   MailOutline,
   MenuFoldOutline,
-  MenuUnfoldOutline,
-  MessageOutline,
-  PayCircleOutline,
+  PhoneOutline,
   PieChartOutline,
-  PrinterOutline,
-  RocketOutline,
+  PlusCircleOutline,
+  QrcodeOutline,
+  QuestionCircleFill,
+  QuestionOutline,
+  RollbackOutline,
+  SafetyOutline,
   ScanOutline,
   SettingOutline,
   ShareAltOutline,
   ShoppingCartOutline,
-  SoundOutline,
+  StarFill,
   StarOutline,
+  SyncOutline,
+  TableOutline,
   TaobaoCircleOutline,
-  TaobaoOutline,
-  TeamOutline,
+  ToTopOutline,
   ToolOutline,
   TrophyOutline,
-  UsbOutline,
   UserOutline,
-  WeiboCircleOutline
+  WechatOutline,
+  WeiboCircleOutline,
+  WeiboOutline,
+  ZhihuOutline,
 ];
